pyproject.toml
[tool.poetry]
name = "root"
version = "0.1"
description = "test task"
authors = [
    "Labanau Nikita <python.dev.back@gmail.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
setuptools = "^67.6.1"
click-plugins = "1.1.1"
pypng = "0.20220715.0"
python-crontab = "3.0.0"
python-dateutil = "2.8.2"
pytz = "2023.3.post1"
requests-toolbelt = "1.0.0"
black = "23.10.1"
celery = "5.3.4"
django = "4.2.6"
django-timezone-field = "6.0.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.0"
drf-spectacular = "0.26.5"
flake8 = "6.1.0"
flower = "2.0.1"
jsonschema = "4.17.3"
more-itertools = "10.1.0"
msgpack = "1.0.7"
mypy-extensions = "1.0.0"
pillow = "10.1.0"
poetry = "1.6.1"
poetry-core = "1.7.0"
poetry-plugin-export = "1.6.0"
psycopg2 = "2.9.9"
pyflakes = "3.1.0"
pyjwt = "2.8.0"
python-dotenv = "1.0.0"
pyyaml = "6.0.1"
qrcode = "7.4.2"
redis = "4.6.0"
django-extensions = "^3.2.3"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py39', 'py310']
include = '\.pyi?$'
skip_magic_trailing_comma = true

[tool.flake8]
exclude = ['tests', 'venv', '.venv', 'deploy', 'tmp_*.py']
ignore = ['W503', 'E203', 'C901']
max-complexity = 10
max-line-length = 120

